pyproject.toml
[project]
name = "capstone-2-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "kagglehub>=0.4.0",
    "numpy>=2.4.1",
    "opencv-python-headless>=4.13.0.90",
    "pillow>=12.1.0",
    "python-multipart>=0.0.21",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]
